+++ src/Data/Extensible/Inclusion.hs
@@ -12,8 +12,8 @@
 ------------------------------------------------------------------------
 module Data.Extensible.Inclusion (
   -- * Membership
-    Position
-  , runPosition
+    Membership
+  , runMembership
   , (∈)()
   , Member(..)
   , Expecting
@@ -49,7 +49,7 @@ type xs ⊆ ys = Include ys xs
 type Include ys = Forall (Member ys)
 
 -- | Reify the inclusion of type level sets.
-inclusion :: forall xs ys. Include ys xs => Position ys :* xs
+inclusion :: forall xs ys. Include ys xs => Membership ys :* xs
 inclusion = generateFor (Proxy :: Proxy (Member ys)) (const membership)
 
 -- | /O(m log n)/ Select some elements.
@@ -69,7 +69,7 @@ spread (UnionAt pos h) = UnionAt (hlookup pos inclusion) h
 {-# INLINE spread #-}
 
 -- | The inverse of 'inclusion'.
-coinclusion :: (Include ys xs, Generate ys) => Nullable (Position xs) :* ys
+coinclusion :: (Include ys xs, Generate ys) => Nullable (Membership xs) :* ys
 coinclusion = flip appEndo (generate (const Null))
   $ hfoldMap getConst'
   $ htabulate (\src dst -> Const' $ Endo $ sectorAt dst `over` const (Eine src))