+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.2.7
+-----------------------------------------------------
+* Added `accessing`
+* Added `decFields` and `decFieldsDeriving`
+* Renamed `Position` to `Membership`
+*
 0.2.6
 -----------------------------------------------------
 * Right-associated `(++)`